v.1.5 A second "supertrim" added: when not using the -a option (so in default mode, stopping after 1st find,) if there's a winning child already found for the winning player, then that one will do. There is no need to keep looking for additional winning children among the siblings of that one, e.g. after that last move by the opponent.
This trim is in the recursive implementation, pending to add it to the
iterative one. This very simple additional trim alone can cut down the
total execution time once again by orders of magnitude. For game-17
(a mate-in-5,) the execution time goes from ~19 mins without the
trim, to < 1 minute! The output can also be massively smaller. In case
of game-17: from a gigantic tree of 190413 nodes to only 5677.

List of updates removed from the README.md and collected into separate file
last_updates.txt.